Support for containers at top level
I found cases where for modularity purposes it was better to have a host in one config file and the containers for this host in other files. For instance if I use a config file for each project but a host is shared by multiple projects it is better to organize it this way.

I still left as valid the case where the host: entry includes containers: as a subentry but I guess I would discourage it.